+++ app/(tabs)/(importrecipe)/_layout.tsx
@@ -1,0 +1,8 @@
+import { Stack } from "expo-router";
+import React from "react";
+
+function Layout() {
+  return <Stack />;
+}
+
+export default Layout;
